fix: clean up product_repo.go
