+++ order-service/internal/service/order_service.go
@@ -14,11 +14,6 @@ const (
 	maxLimit     = 100
 )
 
-const (
-	defaultLimit = 50
-	maxLimit     = 100
-)
-
 type OrderService struct {
 	repo OrderRepository
 }
